# sdks/openapi.yaml
openapi: "3.0.3"
info:
  title: LLMSafeSpace API
  description: |
    LLMSafeSpace is a Kubernetes-first platform for running AI agents securely in isolated sandboxes.
    Every sandbox runs `opencode serve` as a persistent HTTP server with a PVC-backed persistent workspace.
  version: "1.0.0"
  license:
    name: AGPL-3.0-or-later
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: /api/v1
    description: API v1 base path
security:
  - bearerAuth: []
  - apiKeyAuth: []
...
paths:
...
  /admin/settings:
    get:
      tags: [settings]
      summary: Get all admin (instance) settings
      operationId: getAdminSettings
      responses:
        "200":
          description: Admin settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SettingsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/settings/schema:
    get:
      tags: [settings]
      summary: Get admin settings schema
      operationId: getAdminSettingsSchema
      responses:
        "200":
          description: Settings schema
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SettingsSchemaResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/settings/{key}:
    put:
      tags: [settings]
      summary: Set an admin setting
      operationId: setAdminSetting
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [value]
              properties:
                value: {}
      responses:
        "200":
          description: Setting updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  key:
                    type: string
                  value: {}
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /users/me/settings:
    get:
      tags: [settings]
      summary: Get current user settings
      operationId: getUserSettings
      responses:
        "200":
          description: User settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SettingsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /users/me/settings/schema:
    get:
      tags: [settings]
      summary: Get user settings schema
      operationId: getUserSettingsSchema
      responses:
        "200":
          description: User settings schema
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SettingsSchemaResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /users/me/settings/{key}:
    put:
      tags: [settings]
      summary: Set a user setting
      operationId: setUserSetting
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [value]
              properties:
                value: {}
      responses:
        "200":
          description: Setting updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  key:
                    type: string
                  value: {}
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  # --- Credentials (Admin) ---
  /admin/credentials:
    post:
      tags: [credentials]
      summary: Create a credential set
      operationId: createCredentialSet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCredentialSetRequest"
      responses:
        "201":
          description: Credential set created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CredentialSet"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    get:
      tags: [credentials]
      summary: List credential sets
      operationId: listCredentialSets
      responses:
        "200":
          description: Credential set list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CredentialSet"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/credentials/{id}:
    get:
      tags: [credentials]
      summary: Get a credential set
      operationId: getCredentialSet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Credential set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CredentialSet"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [credentials]
      summary: Update a credential set
      operationId: updateCredentialSet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCredentialSetRequest"
      responses:
        "200":
          description: Credential set updated
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [credentials]
      summary: Delete a credential set
      operationId: deleteCredentialSet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Credential set deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Credential set is referenced by workspaces
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/credentials/{id}/default:
    put:
      tags: [credentials]
      summary: Set a credential set as default
      operationId: setDefaultCredentialSet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Default set
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /admin/credentials/rotate-key:
    post:
      tags: [credentials]
      summary: Rotate the credential encryption key
      operationId: rotateCredentialKey
      responses:
        "200":
          description: Rotation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RotateKeyResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  # --- Account ---
...
components:
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from /auth/login or /auth/register
    apiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: "lsp_ prefixed API key in Bearer format (e.g. Bearer lsp_abc123...)"
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    SettingsResponse:
      type: object
      properties:
        settings:
          type: object
          additionalProperties: true
        schemaVersion:
          type: string
    SettingsSchemaResponse:
      type: object
      properties:
        schemaVersion:
          type: string
        settings:
          type: array
          items:
            type: object
            description: Setting definition with key, type, default, description
    CreateCredentialSetRequest:
      type: object
      required: [name, providers]
      properties:
        name:
          type: string
        providers:
          type: object
          description: "Map of provider name to config (e.g. {\"openai\": {\"apiKey\": \"sk-...\", \"baseUrl\": \"...\"}})"
          additionalProperties:
            $ref: "#/components/schemas/ProviderEntry"
        modelAllowlist:
          type: array
          items:
            type: string
        assignedTo:
          description: "\"all\" or array of user IDs"
        isDefault:
          type: boolean
    UpdateCredentialSetRequest:
      type: object
      properties:
        name:
          type: string
        providers:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/ProviderEntry"
        modelAllowlist:
          type: array
          items:
            type: string
        assignedTo:
          description: "\"all\" or array of user IDs"
        isDefault:
          type: boolean
    CredentialSet:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        isDefault:
          type: boolean
        providers:
          type: array
          items:
            type: string
          description: Provider names only (keys never exposed in list responses)
        modelAllowlist:
          type: array
          items:
            type: string
        assignedTo:
          description: "\"all\" or array of user IDs"
        keyVersion:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ProviderEntry:
      type: object
      properties:
        apiKey:
          type: string
        baseUrl:
          type: string
    RotateKeyResult:
      type: object
      properties:
        rotated:
          type: integer
        alreadyCurrent:
          type: integer
        errors:
          type: integer